extract default support imgs automatically on visiting support screen (if support img folder doesn't exist)
