+++ app/src/main/res/menu/support_extract_menu.xml
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/action_support_extract_defaults"
+        android:icon="@drawable/ic_zip"
+        android:title="Extract Default Support Images"
+        app:showAsAction="never" />
+</menu>
